Change file directory and packaging
Chang classes directory from src/main/froggy to
src/main/java/froggy and change package to froggy
