+++ build.gradle
@@ -28,11 +28,11 @@ test {
 }
 
 application {
-    mainClass.set("seedu.duke.Duke")
+    mainClass.set("froggy.Froggy")
 }
 
 shadowJar {
-    archiveBaseName = "duke"
+    archiveBaseName = "froggy"
     archiveClassifier = null
 }
