+++ src/main/java/froggy/AddEventCommand.java
@@ -1,4 +1,4 @@
-package main.froggy;
+package froggy;
 
 public class AddEventCommand extends Command {
