+++ src/main/java/froggy/Event.java
@@ -1,4 +1,4 @@
-package main.froggy;
+package froggy;
 
 public class Event extends Task {
