+++ src/main/java/froggy/ExitCommand.java
@@ -1,4 +1,4 @@
-package main.froggy;
+package froggy;
 
 public class ExitCommand extends Command {
