+++ src/main/java/froggy/Froggy.java
@@ -1,4 +1,4 @@
-package main.froggy;
+package froggy;
 
 public class Froggy {
     private static final String FILE_PATH = "./data/taskList.txt";
