+++ src/main/java/froggy/ListCommand.java
@@ -1,4 +1,4 @@
-package main.froggy;
+package froggy;
 
 public class ListCommand extends Command{
