+++ src/main/java/froggy/Storage.java
@@ -1,4 +1,4 @@
-package main.froggy;
+package froggy;
 import java.io.BufferedReader;
 import java.io.BufferedWriter;
 import java.io.File;
