+++ src/main/java/froggy/TaskList.java
@@ -1,4 +1,4 @@
-package main.froggy;
+package froggy;
 
 import java.util.ArrayList;
 import java.util.List;
