+++ src/main/java/froggy/Todo.java
@@ -1,4 +1,4 @@
-package main.froggy;
+package froggy;
 
 public class Todo extends Task{
