+++ src/main/java/froggy/Ui.java
@@ -1,4 +1,4 @@
-package main.froggy;
+package froggy;
 
 import java.util.Scanner;
